api/requirements.txt
flask
dramatiq[redis]
dramatiq-abort[redis]
requests
pyyaml
tqdm
django-environ
python-slugify

pandas
Pillow<10.0
scipy
scikit-learn
matplotlib
seaborn
opencv-python

kornia==0.7.*
torch==1.12.1+cu116
torchvision==0.13.1+cu116
torchaudio==0.12.1
--extra-index-url https://download.pytorch.org/whl/cu116
